refactor(updateSelOpts, eventManager) :: 파일 분리
