+++ src/basic/main.basic.js
@@ -1,6 +1,7 @@
 import calculateCart from './calculateCart';
-import { createOption } from './createElements';
+import { setLuckySale, setSuggestSale } from './eventManager';
 import renderHome from './renderHome';
+import updateSelOpts from './updateSelOpts';
 
 let prodList, sel, addBtn, cartsDiv, sumDiv, stockInfoDiv;
 let lastSel;
@@ -27,43 +28,8 @@ function main() {
 
   calculateCart({ prodList, sumDiv, cartsDiv, stockInfoDiv });
 
-  setTimeout(function () {
-    setInterval(function () {
-      let luckyItem = prodList[Math.floor(Math.random() * prodList.length)];
-      if (Math.random() < 0.3 && luckyItem.q > 0) {
-        luckyItem.val = Math.round(luckyItem.val * 0.8);
-        alert('번개세일! ' + luckyItem.name + '이(가) 20% 할인 중입니다!');
-        updateSelOpts(sel, prodList);
-      }
-    }, 30000);
-  }, Math.random() * 10000);
-
-  setTimeout(function () {
-    setInterval(function () {
-      if (lastSel) {
-        let suggest = prodList.find(function (item) {
-          return item.id !== lastSel && item.q > 0;
-        });
-        if (suggest) {
-          alert(suggest.name + '은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!');
-          suggest.val = Math.round(suggest.val * 0.95);
-          updateSelOpts(sel, prodList);
-        }
-      }
-    }, 60000);
-  }, Math.random() * 20000);
-}
-
-function updateSelOpts(sel, prodList) {
-  if (!sel) return null;
-
-  sel.innerHTML = '';
-  prodList.forEach((item) => {
-    const { id, name, val, q } = item;
-    const opt = createOption({ val: id, text: `${name} - ${val}원`, disabled: q === 0 });
-
-    sel.appendChild(opt);
-  });
+  setLuckySale(prodList, sel);
+  setSuggestSale(prodList, sel, lastSel);
 }
 
 main();
